+++ src/global.h
@@ -61,10 +61,10 @@
 
 /* structure representing a file or directory */
 struct dir {
-  int64_t size, asize;
+  uint64_t size, asize;
   uint64_t ino, dev;
   struct dir *parent, *next, *prev, *sub, *hlnk;
-  int items;
+  uint64_t items;
   unsigned char flags;
   char name[3]; /* must be large enough to hold ".." */
 };
